requirements.txt
numpy >= 1.8.2
sympy >= 1.1.1
antlr4-python3-runtime >= 4.7
setuptools
Jinja2 >= 2.10
typing
astropy == 2.0.3
odetoolbox >= 2.0
